Move di files
