+++ lib/src/structure/di/categories.dart
@@ -1,7 +1,7 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../app_state/game_config.dart';
 import '../../fs_interface/di/fs_watcher.dart';
+import '../../storage/di/game_config.dart';
 import '../entity/mod_category.dart';
 import '../usecase/collect_categories.dart';
