+++ lib/src/ui/route/home_shell.dart
@@ -9,15 +9,15 @@ import 'package:protocol_handler/protocol_handler.dart';
 import 'package:window_manager/window_manager.dart';
 
 import '../../../l10n/app_localizations.dart';
-import '../../app_state/current_target_game.dart';
-import '../../app_state/exe_arg.dart';
-import '../../app_state/game_config.dart';
-import '../../app_state/games_list.dart';
-import '../../app_state/run_together.dart';
-import '../../app_state/separate_run_override.dart';
-import '../../app_state/window_size.dart';
 import '../../app_version/di/is_outdated.dart';
 import '../../fs_interface/helper/path_op_string.dart';
+import '../../storage/di/current_target_game.dart';
+import '../../storage/di/exe_arg.dart';
+import '../../storage/di/game_config.dart';
+import '../../storage/di/games_list.dart';
+import '../../storage/di/run_together.dart';
+import '../../storage/di/separate_run_override.dart';
+import '../../storage/di/window_size.dart';
 import '../../structure/di/categories.dart';
 import '../constants.dart';
 import '../util/display_infobar.dart';
