+++ lib/src/ui/route/loading.dart
@@ -4,7 +4,7 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../app_state/games_list.dart';
+import '../../storage/di/games_list.dart';
 import '../../storage/di/storage.dart';
 import '../constants.dart';
 import '../widget/appbar.dart';
