+++ lib/src/ui/route/setting.dart
@@ -10,22 +10,22 @@ import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:smooth_scroll_multiplatform/smooth_scroll_multiplatform.dart';
 
-import '../../app_state/card_color.dart';
-import '../../app_state/column_strategy.dart';
-import '../../app_state/dark_mode.dart';
-import '../../app_state/display_enabled_mods_first.dart';
-import '../../app_state/folder_icon.dart';
-import '../../app_state/game_config.dart';
-import '../../app_state/move_on_drag.dart';
-import '../../app_state/run_together.dart';
-import '../../app_state/separate_run_override.dart';
-import '../../app_state/use_paimon.dart';
-import '../../app_state/value_settable.dart';
 import '../../app_version/di/current_version.dart';
 import '../../app_version/di/is_outdated.dart';
 import '../../app_version/di/remote_version.dart';
 import '../../fs_interface/di/fs_interface.dart';
 import '../../fs_interface/entity/setting_data.dart';
+import '../../storage/di/card_color.dart';
+import '../../storage/di/column_strategy.dart';
+import '../../storage/di/dark_mode.dart';
+import '../../storage/di/display_enabled_mods_first.dart';
+import '../../storage/di/folder_icon.dart';
+import '../../storage/di/game_config.dart';
+import '../../storage/di/move_on_drag.dart';
+import '../../storage/di/run_together.dart';
+import '../../storage/di/separate_run_override.dart';
+import '../../storage/di/use_paimon.dart';
+import '../../storage/di/value_settable.dart';
 import '../../storage/domain/entity/game_config.dart';
 import '../../storage/domain/usecase/card_color.dart';
 import '../constants.dart';
