+++ lib/src/ui/widget/protocol_handler.dart
@@ -7,10 +7,10 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../app_state/exe_arg.dart';
 import '../../nahida/di/nahida_download_queue.dart';
 import '../../nahida/di/nahida_store.dart';
 import '../../nahida/domain/entity/nahida_element.dart';
+import '../../storage/di/exe_arg.dart';
 import '../../structure/di/categories.dart';
 import '../../structure/entity/mod_category.dart';
 import '../util/display_infobar.dart';
