+++ test/src/ui/route/welcome_test.dart
@@ -1,6 +1,6 @@
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:genshin_mod_manager/src/app_state/current_target_game.dart';
+import 'package:genshin_mod_manager/src/storage/di/current_target_game.dart';
 import 'package:genshin_mod_manager/src/ui/route/welcome.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
